requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
sqlparse==0.5.1
tzdata==2024.1
pip==24.1.2
setuptools==65.5.0
django-cors-headers==4.4.0